Split test into folders
