+++ OakChan.Tests/Unit/Controllers/BoardControllerTests.cs
@@ -1,5 +1,4 @@
 ﻿using AutoMapper;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
 using Microsoft.Extensions.Logging;
@@ -8,11 +7,11 @@ using NUnit.Framework;
 using OakChan.Controllers;
 using OakChan.Services;
 using OakChan.Services.DTO;
-using OakChan.Tests.Base;
+using OakChan.Tests.Integration.Base;
 using OakChan.ViewModels;
 using System.Threading.Tasks;
 
-namespace OakChan.Tests.Controllers
+namespace OakChan.Tests.Unit.Controllers
 {
     public class BoardControllerTests : ControllerTestsBase
     {
