+++ OakChan.Tests/Unit/Markup/QuoteTagTests.cs
@@ -3,7 +3,7 @@ using OakChan.Markup.Tags;
 using System;
 using System.Threading.Tasks;
 
-namespace OakChan.Tests.Markup
+namespace OakChan.Tests.Unit.Markup
 {
     public class QuoteTagTests
     {
